refactor(character): extract data layer into useHeroDetail hook (native)
Move every fetch, piece of loaded state, and derived value for the
character detail screen out of app/character/[id].tsx and into a new
platform-neutral hook, src/hooks/useHeroDetail.ts. The native view now
consumes it by identical names, so its render body is unchanged while
the data layer is owned in one place.

This is step 1 of de-duplicating the character screen pair: the same
hook will back app/character/[id].web.tsx in a follow-up, deleting the
copy-pasted fetch/state logic that had drifted between the two ~2.5-4k
line files. Haptics calls in the hook are guarded with Platform.OS so
the hook is safe to share with web.

No behavior change on native. typecheck clean; 363 tests pass.
